Tinker Tinkering, Add ownership to listings, Manage Listings Page and User Authorization

+++ app/Http/Controllers/ListingController.php
@@ -38,6 +38,8 @@ class ListingController extends Controller
         return view('listings.create');
     }
 
+
+
     //store new listing data
     public function store(Request $request)
     {
@@ -55,6 +57,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->user()->id;
+
 
         Listing::create($formFields);
 
@@ -70,6 +74,12 @@ class ListingController extends Controller
     //update listing data
     public function update(Request $request, Listing $listing)
     {
+
+        //Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'company' => ['required'],
             'title' => 'required',
@@ -93,7 +103,19 @@ class ListingController extends Controller
     //Delete Listing
     public function delete(Listing $listing)
     {
-        $listing->delete() ;
+        //Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
+        $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+
+    //manage listings
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
